requirements.txt
streamlit
sentence-transformers
PyPDF2
python-docx
numpy
pandas
regex